Rename app-information to member-reminders
Discussion:
https://github.com/HedvigInsurance/android/pull/1720#pullrequestreview-1551894124
