+++ app/app/build.gradle.kts
@@ -196,7 +196,6 @@ dependencies {
   implementation(projects.apolloDi)
   implementation(projects.apolloGiraffePublic)
   implementation(projects.apolloOctopusPublic)
-  implementation(projects.appInformationPublic)
   implementation(projects.audioPlayer)
   implementation(projects.authAndroid)
   implementation(projects.authCore)
@@ -225,6 +224,7 @@ dependencies {
   implementation(projects.hanalyticsFeatureFlagsPublic)
   implementation(projects.languageCore)
   implementation(projects.marketCore)
+  implementation(projects.memberRemindersPublic)
   implementation(projects.navigationActivity)
   implementation(projects.navigationComposeTyped)
   implementation(projects.navigationCore)
