+++ app/member-reminders/member-reminders-public/build.gradle.kts
