+++ app/member-reminders/member-reminders-public/src/main/AndroidManifest.xml
