+++ app/member-reminders/member-reminders-public/src/main/graphql/com/hedvig/android/memberreminders/graphql/QueryGetPayinMethodStatus.graphql
