+++ app/member-reminders/member-reminders-public/src/main/kotlin/com/hedvig/android/memberreminders/EnableNotificationsReminderManager.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.appinformation
+package com.hedvig.android.memberreminders
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
