+++ app/member-reminders/member-reminders-public/src/main/kotlin/com/hedvig/android/memberreminders/GetUpcomingRenewalRemindersUseCase.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.appinformation
+package com.hedvig.android.memberreminders
 
 import arrow.core.Either
 import arrow.core.NonEmptyList
