+++ app/member-reminders/member-reminders-public/src/test/kotlin/com/hedvig/android/memberreminders/EnableNotificationsReminderManagerTest.kt
@@ -1,4 +1,4 @@
-package com.hedvig.android.appinformation
+package com.hedvig.android.memberreminders
 
 import assertk.assertThat
 import assertk.assertions.isFalse
